# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Curiosity Box API
  description: |
    A safe learning platform where kids ask questions within parent-approved topics, powered by Claude AI.

    ## Authentication
    The API uses token-based authentication. Protected endpoints require an `Authorization` header with the format:
    ```
    Authorization: Token <your-token-here>
    ```

    ## Rate Limiting
    **Note:** Rate limiting is recommended for production but not currently implemented.

    **Recommended limits:**
    - Anonymous users: 10 requests/minute
    - Authenticated users: 100 requests/minute
    - AI question generation: 20 requests/minute per child

    When rate limited, the API will return a `429 Too Many Requests` response with a `Retry-After` header.

    ## Pagination
    All list endpoints support pagination using query parameters:
    - `page` - Page number (default: 1)
    - `page_size` - Items per page (max: 20, default: 20)

    Paginated responses include:
    - `count` - Total number of items
    - `next` - URL to next page (null if last page)
    - `previous` - URL to previous page (null if first page)
    - `results` - Array of items for current page

  version: 1.0.0
  contact:
    name: Curiosity Box API Support
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.curiositybox.example.com
    description: Production server
...
paths:
...
  /api/children/{id}/:
    get:
      tags:
        - Children
      summary: Get child details
      description: Get details for a specific child including their topic access. Requires authentication.
      operationId: getChild
      security:
        - TokenAuth: []
      parameters:
        - $ref: '#/components/parameters/ChildIdParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Child'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/children/{id}/topics/enable/:
    post:
      tags:
        - Children
      summary: Enable a topic for a child
      description: Grant a child access to ask questions about a specific topic. Requires authentication.
      operationId: enableTopic
      security:
        - TokenAuth: []
      parameters:
        - $ref: '#/components/parameters/ChildIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - topic_slug
              properties:
                topic_slug:
                  type: string
                  example: animals
      responses:
        '200':
          description: Topic enabled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Topic "Animals" enabled for Emma
                  created:
                    type: boolean
                    description: True if newly created, false if already existed
                    example: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/children/{id}/topics/disable/:
    post:
      tags:
        - Children
      summary: Disable a topic for a child
      description: Remove a child's access to ask questions about a specific topic. Requires authentication.
      operationId: disableTopic
      security:
        - TokenAuth: []
      parameters:
        - $ref: '#/components/parameters/ChildIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - topic_slug
              properties:
                topic_slug:
                  type: string
                  example: space
      responses:
        '200':
          description: Topic disabled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Topic access removed
                  deleted:
                    type: boolean
                    description: True if topic access was removed, false if didn't exist
                    example: true
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Questions Endpoints
...
components:
  securitySchemes:
    TokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Token-based authentication. Format - `Token <your-token>`
  parameters:
...
    ChildIdParam:
      name: id
      in: path
      required: true
      description: Child ID
      schema:
        type: integer
      example: 1
  schemas:
...
    TopicCategory:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Animals
        slug:
          type: string
          example: animals
        description:
          type: string
          example: Learn about wildlife, pets, and creatures from around the world
        icon:
          type: string
          example: 🦁
        recommended_min_age:
          type: integer
          minimum: 3
          example: 5
        is_active:
          type: boolean
          example: true
    ChildTopicAccess:
      type: object
      properties:
        id:
          type: integer
          example: 1
        topic:
          $ref: '#/components/schemas/TopicCategory'
        enabled_at:
          type: string
          format: date-time
          example: '2026-01-11T10:30:00Z'
    Child:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Emma
        age:
          type: integer
          minimum: 3
          maximum: 18
          example: 8
        reading_level:
          type: string
          enum:
            - early
            - intermediate
            - advanced
          example: intermediate
        avatar_color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          example: '#4A90E2'
        topic_access:
          type: array
          items:
            $ref: '#/components/schemas/ChildTopicAccess'
        enabled_topics:
          type: array
          items:
            $ref: '#/components/schemas/TopicCategory'
...
  responses:
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: Authentication credentials were not provided.
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: Not found.
...